updated lh side bar, all-caps subheading typography
