+++ src/components/shared/Footer/Footer.tsx
@@ -14,8 +14,6 @@ import UcalgaryLogo from 'assets/images/University-Of-Calgary-Logo.png';
 import AmcJoule from 'assets/images/amc-joule.png';
 import './Footer.scss';
 
-const lancetId = "https://www.thelancet.com/journals/laninf/article/PIIS1473-3099(20)30631-9/fulltext#%20"
-
 type UpdatedAtProps = {
   updatedAt: string;
 }
@@ -33,9 +31,6 @@ const renderDesktopFooter = (updatedAt: string) => (
       <div className="footer__visible-section row justify-content-between d-flex align-items-center text-center">
         <div className="col-2">
           <UpdatedAt updatedAt={updatedAt}/>
-        </div>
-        <div className="col-2">
-            <Citation/>
         </div>
       </div>
       <div className="row justify-content-center mt-3">
@@ -67,7 +62,6 @@ const renderMobileFooter = (updatedAt: string) => (
   <footer className={'container-fluid mx-0 footer'}>
       <div className="footer__visible-section row d-flex p-2">
           <UpdatedAt updatedAt={updatedAt}/>
-          <Citation/>
       </div>
       <div className="row justify-content-center mt-5">
         <div className="col-10 ">
@@ -108,15 +102,6 @@ const UpdatedAt = ({updatedAt}: UpdatedAtProps) => {
   ): null
 }
 
-const Citation = () => (
-  <span className="text-right">
-    {Translate("Footer", ["CiteAs"])}
-    <a href={lancetId} target="__blank" rel="noopener noreferrer" className="cite-link">
-      <i>{Translate("Footer", ["LancetInfDis"])}</i> {Translate("Footer", ["Article"])}
-    </a> 
-  </span>
-)
-
 const WhoDisclaimer = () => (
   <small className="whoDisclaimer">
     {Translate('WhoSerotrackAndPartnersDisclaimerSmall')}
@@ -149,7 +134,7 @@ const PageLinks = () => (
       Object.keys(PAGE_HASHES).map((page) => {
         return(
         <div className="col mx-2">
-          <h3 className="row">{Translate(page)}</h3>
+          <p className="row subheading">{Translate(page)}</p>
           {
           Object.keys(PAGE_HASHES[page]).map(h => (
           <HashLink to={`${withLocaleUrl(page)}#${h}`} className="row mt-3 footer__link">
