+++ src/sass/_typography.scss
@@ -33,3 +33,12 @@ h3 {
 h4 {
   font-size: 20.2px;
 }
+
+.subheading {
+  font-size: 14px;
+  font-weight: 600;
+}
+
+.extra-bold {
+  font-weight: 800;
+}
