+++ src/sass/app.scss
@@ -67,6 +67,7 @@
 
 .download-data-btn.ui.button {
     background-color: $stone-cold-steve-austin;
+    width: 100% !important;
     a {
         color: white;
     }
@@ -83,7 +84,11 @@
 }
 
 .cite-link {
-    color: #55A6BA;
+    color: #0066CC;
+    cursor: pointer;
+    &:hover {
+        color: #2f00a1;
+    }
 }
 
 .italics{
@@ -341,9 +346,10 @@
 }
 
 .link {
-    color: #0066CC;
-    cursor: pointer;
-    &:hover {
-        color: #2f00a1;
-    }
+    text-decoration: underline;
+    font-weight: 900;
+}
+
+.full-width-button {
+    width: 100%;
 }
