+++ src/utils/translate/en.json
@@ -2,7 +2,7 @@
   ",": ",",
   "95%ConfidenceInterval": "95% Confidence Interval",
   "95%ConfidenceIntervalTooltip": "Wider error bars indicate greater uncertainty in the pooled seroprevalence estimate.",
-  "About": "About",
+  "About": "ABOUT",
   "AboutPage": {
     "AboutSection": {
       "Covid19ImmunityTaskForce": "COVID-19 Immunity Task Force.",
@@ -79,7 +79,7 @@
     "Title": "Analysis Methods",
     "Tooltip": "The aggregation technique used is fixed-effects, inverse-variance-weighted meta-analysis of prevalence, with the Freeman-Tukey double arcsin transform used to stabilize variance. Records over sample size MIN_DENOMINATOR considered."
   },
-  "Analyze": "Analyze",
+  "Analyze": "ANALYZE",
   "AntibodyTarget": "Antibody Target",
   "AntibodyTargetOptions": {
     "EnvelopeProtein": "Envelope protein",
@@ -135,7 +135,7 @@
   },
   "COVIDImmunityTaskForce": "COVID-19 Immunity Task Force",
   "Dashboard": "Dashboard",
-  "Data": "Data",
+  "Data": "DATA",
   "DatabaseUptoDateTo": "Database up to date to",
   "DataDictionary": "Data dictionary",
   "DataDictionaryText": {
@@ -229,13 +229,13 @@
   },
   "EstimateRange": "Estimate Range",
   "ExecutiveDirector": "Executive Director",
-  "Explore": "Explore",
+  "Explore": "EXPLORE",
   "ExploreBlurb": {
     "FirstParagraph": "SeroTracker is a dashboard and data platform for SARS-CoV-2 serosurveys.",
     "SecondParagraph": "We conduct an ongoing systematic review to track serosurveys (antibody testing-based surveillance efforts) around the world and visualize findings on this dashboard.",
     "ThirdParagraph": "Antibodies against SARS-CoV-2, the virus that causes COVID-19, indicate previous infection or vaccination. "
   },
-  "Filter": "Filter",
+  "Filter": "FILTERS",
   "FilterTooltip": "Filter the SeroTracker dataset based on geography, study, demographics, and test. Only studies matching all of the listed filters will be included.",
   "For": "For",
   "Footer": {
@@ -395,7 +395,7 @@
   "Paper": "Paper",
   "PaperText": "Our first manuscript, describing lessons from serosurveys up to May 1, can be found on",
   "PaperTextTwo": "More information about our dashboard can be found in",
-  "Partnerships": "Partnerships",
+  "Partnerships": "PARTNERSHIPS",
   "PartnershipsList": {
     "Canada": "Canada (CITF)"
   },
@@ -450,7 +450,7 @@
     "MediaMentions": "Worldwide media publications featuring SeroTracker",
     "PrivateSectorReports": "Reports on COVID-19 screening and testing practices across Canada’s private sector"
   },
-  "Publications": "Publications",
+  "Publications": "PUBLICATIONS",
   "Published": "Published",
   "DataTable": "Data Table",
   "RegionalEstimateRange": "Regional Estimate Range",
@@ -553,7 +553,9 @@
   "SublocalStudyDetails": "Sublocal Study",
   "SubmitSource": "Submit source",
   "Summary": "Summary",
-  "SummaryStatistics": "Summary Statistics",
+  "SummaryStatistics": "KEY INSIGHTS",
+  "SummaryStatisticsLink1": "Read biweekly insights",
+  "SummaryStatisticsLink2": "here",
   "Team": "Team",
   "TeamLead": "Team Leads",
   "TemporaryMaintenance": "Temporary Maintenance",
@@ -602,7 +604,7 @@
   "ViewPartnership": "View Partnership",
   "ViewFile": "View File",
   "ViewHere": "View Here",
-  "Welcome": "Welcome to SeroTracker",
+  "Welcome": "ABOUT US",
   "WhatWeDoText": {
     "FirstParagraph": "At SeroTracker, we collect data and information on SARS-CoV-2 seroprevalence studies by conducting a live systematic review and meta-analysis of published and unpublished data on  seroprevalence. Data is extracted for each source's primary estimate and certain relevant subgroups. Full methods are available in our protocol.",
     "SecondParagraph": "We produce data for both academic and public use on our dashboard. We aim to update our dashboard weekly. For further information beyond the following high-level summary, please see our study protocol and supplementary methods for a detailed explanation of our research processes."
